Add files via upload
This is a simple dashboard with some measures to calculate sales, budget, YOY, transactions and to have a glimpse of how you can handle a many to many relationship in Power BI or connect different tables to your model. There's a lot of work to do with the design, measures and model but hopefully you can have some inspiration.