Add files via upload
This is another simple dashboard for global sales and diferent products. The main measures are month related, but you can adapt it, for weeks, years and so on. It's another model with several tables where we have to create some relationships in order our model to work, so maybe you can get some ideas for your data.